+++ CHANGELOG.md
@@ -2,5 +2,11 @@
 
 All notable changes will be documented in this file
 
+## v0.9.2 - 2017-10-30
+- Cast currency-code to currency instance on amount instantiation
+
+## v0.9.1 - 2017-10-30
+- Add Laravel 5.5 auto discover support
+
 ## v0.9.0 - 2017-10-30
 - Initial release